pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "noteexpander"
version = "0.1.0"
description = "Discord bot + local dashboard that turns brief ideas into AI-researched reports"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "openai-whisper",
    "litellm",
    "tavily-python",
    "fastapi>=0.135.0",
    "uvicorn[standard]",
    "jinja2",
    "markdown",
    "bleach",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite",
    "pyyaml",
    "python-dotenv",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "pytest", "pytest-asyncio"]

[project.scripts]
noteexpander = "main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
